papr_be/requirements.txt
awsebcli==3.10.4
beautifulsoup4==4.6.0
blessed==1.14.2
boto3==1.4.5
botocore==1.5.83
cement==2.8.2
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.7
cssselect==1.0.1
Django==1.11.4
djangorestframework==3.6.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.6
jieba3k==0.35.1
jmespath==0.9.3
jsonfield
lxml==4.0.0
nltk==3.2.4
olefile==0.44
pathspec==0.5.0
Pillow==4.2.1
psycopg2==2.7.3
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.9.1
requests-file==1.4.2
s3transfer==0.1.10
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
termcolor==1.1.0
tldextract==2.1.0
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
websocket-client==0.44.0
